Consolidate shared/data and shared/domain into app/
Move app-level data and domain classes to app/ package:
- AppDetailsProvider (all platforms) → app/data/
- AndroidContextHolder, DesktopStoragePaths → app/data/
- Exporter → app/data/
- AppDetails → app/domain/model/
- LoadAppDetailsUseCase → app/domain/usecase/

Also fix misplaced tests - tests in shared/data/ and shared/domain/
were actually testing feature classes, moved to correct locations:
- MemoMapper/MemosApi/MemosRepositoryImpl tests → feature/memos/data/
- HabitParser/HabitContentBuilder tests → feature/habits/domain/
- MemoEntityMapper test → feature/memos/data/local/
