+++ androidApp/src/main/kotlin/space/be1ski/vibits/android/VibitsApplication.kt
@@ -1,7 +1,7 @@
 package space.be1ski.vibits.android
 
 import android.app.Application
-import space.be1ski.vibits.shared.data.local.AndroidContextHolder
+import space.be1ski.vibits.shared.app.data.AndroidContextHolder
 
 /**
  * Android Application that initializes shared context.
