+++ shared/src/commonMain/kotlin/space/be1ski/vibits/shared/app/data/AppDetailsProvider.kt
@@ -1,0 +1,13 @@
+package space.be1ski.vibits.shared.app.data
+
+import space.be1ski.vibits.shared.app.domain.model.AppDetails
+
+/**
+ * Provides application details for settings and diagnostics.
+ */
+expect class AppDetailsProvider() {
+  /**
+   * Returns app details including version, environment, and storage paths.
+   */
+  fun load(): AppDetails
+}
