--- shared/src/commonMain/kotlin/space/be1ski/vibits/shared/domain/usecase/LoadAppDetailsUseCase.kt
@@ -1,15 +1,0 @@
-package space.be1ski.vibits.shared.domain.usecase
-
-import dev.zacsweers.metro.Inject
-import space.be1ski.vibits.shared.data.local.AppDetailsProvider
-import space.be1ski.vibits.shared.domain.model.app.AppDetails
-
-/**
- * Loads app details for settings screen.
- */
-@Inject
-class LoadAppDetailsUseCase(
-  private val appDetailsProvider: AppDetailsProvider,
-) {
-  operator fun invoke(): AppDetails = appDetailsProvider.load()
-}
