+++ shared/src/commonMain/kotlin/space/be1ski/vibits/shared/feature/settings/presentation/SettingsFeature.kt
@@ -1,6 +1,6 @@
 package space.be1ski.vibits.shared.feature.settings.presentation
 
-import space.be1ski.vibits.shared.domain.model.app.AppDetails
+import space.be1ski.vibits.shared.app.domain.model.AppDetails
 import space.be1ski.vibits.shared.feature.mode.domain.model.AppMode
 import space.be1ski.vibits.shared.feature.settings.domain.model.AppLanguage
 import space.be1ski.vibits.shared.feature.settings.domain.model.AppTheme
