+++ shared/src/desktopMain/kotlin/space/be1ski/vibits/shared/app/data/AppDetailsProvider.kt
@@ -1,0 +1,21 @@
+package space.be1ski.vibits.shared.app.data
+
+import space.be1ski.vibits.shared.app.domain.model.AppDetails
+import java.nio.file.Paths
+
+/**
+ * Desktop implementation.
+ */
+actual class AppDetailsProvider {
+  actual fun load(): AppDetails {
+    val home = System.getProperty("user.home")
+    val offlinePath = Paths.get(home, "Documents", "Vibits", "memos.json").toString()
+    return AppDetails(
+      version = DesktopStoragePaths.appVersion(),
+      environment = DesktopStoragePaths.environmentLabel(),
+      credentialsStore = "Preferences(${DesktopStoragePaths.preferencesNode()})",
+      memosDatabase = DesktopStoragePaths.databasePath(),
+      offlineStorage = offlinePath,
+    )
+  }
+}
