+++ shared/src/desktopMain/kotlin/space/be1ski/vibits/shared/feature/settings/data/PreferencesStore.kt
@@ -1,6 +1,6 @@
 package space.be1ski.vibits.shared.feature.settings.data
 
-import space.be1ski.vibits.shared.data.local.DesktopStoragePaths
+import space.be1ski.vibits.shared.app.data.DesktopStoragePaths
 import java.util.prefs.Preferences
 
 /**
