+++ shared/src/wasmJsMain/kotlin/space/be1ski/vibits/shared/app/data/AppDetailsProvider.kt
@@ -1,0 +1,17 @@
+package space.be1ski.vibits.shared.app.data
+
+import space.be1ski.vibits.shared.app.domain.model.AppDetails
+
+/**
+ * Web implementation.
+ */
+actual class AppDetailsProvider {
+  actual fun load(): AppDetails =
+    AppDetails(
+      version = "web",
+      environment = "web",
+      credentialsStore = "in-memory",
+      memosDatabase = "in-memory",
+      offlineStorage = "localStorage",
+    )
+}
